# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MH GenAI Engine API
  version: 0.4.0
  description: |
    Versioned API for the MH GenAI Engine. Provides OER-grounded RAG,
    template-driven generation, guardrails, and telemetry.
    Authentication/authorization TBD (e.g. API key header).
    Changelog: see docs/api/CHANGELOG_API.md. Deferred items and open questions: docs/api/API_DEFERRED_AND_NOTES.md.

servers:
  - url: https://dev.mh-genai-engine.com/v0
    description: Dev environment

paths:
  /generate/lesson-outline:
    post:
      summary: Generate a lesson outline grounded in OER content
      description: |
        Retrieves grounded passages from the vector store, then calls the configured LLM to produce
        structured JSON (`outline`, `keyConcepts`, etc.). **Citations** are added by the Engine: one
        object per retrieved chunk, same order as passages in the model context. Index ``i`` in
        ``citations[i]`` matches ``### Passage [i]`` and valid ``<grounded ref="i">…</grounded>``
        spans in string fields (including ``slideOutline`` when ``contentType`` is ``ppt``).
        Requires ``DATABASE_URL``, embeddings, indexed content, and a configured LLM (e.g. Anthropic).
        On LLM/parse failure the server may respond with **502**.
      operationId: generateLessonOutline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonOutlineRequest"
      responses:
        "200":
          description: Successful generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonOutlineResponse"
        "400":
          description: Bad request (e.g. validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized (auth TBD)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Bad gateway — lesson outline generation failed (e.g. LLM returned invalid JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /generate/assessment-transform:
    post:
      summary: Transform an MCQ into an open-ended question + rubric
      description: |
        Workflow 2 scope is being redesigned (Assessment Generation Tool).
        This endpoint and its schemas are retained as placeholders pending
        the new Workflow 2 API contract.
      operationId: transformAssessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentTransformRequest"
      responses:
        "200":
          description: Successful transformation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentTransformResponse"
        "400":
          description: Bad request (e.g. validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized (auth TBD)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /templates/{workflow}:
    get:
      summary: List templates for a workflow
      description: |
        Returns Engine-approved **prompt template** ids for one workflow. Each item's ``id`` is a
        stable kebab-case string: use ``lesson-outline`` ids as ``template`` on
        ``POST /generate/lesson-outline``. Assessment-transform ids are reserved for Workflow 2 when
        a template field is added there.
      operationId: listTemplatesByWorkflow
      parameters:
        - name: workflow
          in: path
          required: true
          description: Workflow slug
          schema:
            type: string
            enum: [lesson-outline, assessment-transform]
      responses:
        "200":
          description: Template list for the requested workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateList"
        "401":
          description: Unauthorized (auth TBD)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Unknown workflow path value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    LessonOutlineRequest:
      type: object
      properties:
        book:
          type: string
          description: Book identifier (pre-selected for POC; only one book available)
        chapter:
          type: string
          nullable: true
          description: |
            Chapter identifier from the content catalog when chunks have TOC-backed
            metadata. Omit or set null when scoping by chapter does not apply (e.g. PDFs
            with no table of contents). Empty string is also accepted and behaves like
            no chapter filter at retrieval time.
        section:
          type: string
          description: Section identifier within the chapter (optional content scoping)
        subSection:
          type: string
          description: Sub-section identifier within the section (optional content scoping)
        learningObjective:
          type: string
          description: Single learning objective statement provided by the instructor
        contentType:
          type: string
          enum: [lecture_notes, ppt]
          description: Output format — lecture notes (time-structured) or ppt (slide-structured outline)
        count:
          type: integer
          description: |
            Target output size. Interpretation depends on contentType:
            when lecture_notes, this is duration in minutes;
            when ppt, this is number of slides.
        audienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
          description: Target audience level
        regeneratedResponse:
          type: boolean
          default: false
          description: |
            Logging-only flag. Set to true when the user is regenerating
            (re-requesting with updated parameters). Does not change
            generation behavior.
        template:
          type: string
          default: default
          enum: [default, lecture-scaffold-one-shot]
          description: |
            Prompt template variant (kebab-case). Must match an ``id`` from
            ``GET /templates/lesson-outline``. The Engine maps this to the system prompt;
            clients do not send prompt text.
      required:
        - learningObjective
        - contentType
        - count
        - audienceLevel

    LessonOutlineResponse:
      type: object
      properties:
        outline:
          type: string
          description: |
            Main lesson or lecture outline text, aligned to objectives. May include
            ``<grounded ref="i">verbatim excerpt</grounded>`` for UI linking to ``citations[i]``.
        keyConcepts:
          type: array
          items: { type: string }
          description: Key concepts to cover (optional)
        misconceptions:
          type: array
          items: { type: string }
          description: Common misconceptions to address (optional)
        checksForUnderstanding:
          type: array
          items: { type: string }
          description: Checks aligned to objectives (optional)
        activityIdeas:
          type: array
          items: { type: string }
          description: Suggested in-class activity ideas, time-boxed (optional)
        slideOutline:
          type: string
          description: |
            Slide-by-slide or speaker-note scaffolding when ``contentType`` is ``ppt``; use the same
            ``<grounded ref="i">…</grounded>`` convention as ``outline`` (not ad-hoc ``[Ref i]`` markers).
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
          description: |
            One entry per retrieved chunk, same order as passages in the model context.
            Index ``i`` matches ``<grounded ref="i">``. ``snippet`` is a short preview (50 chars max).

    AssessmentTransformRequest:
      type: object
      properties:
        question:
          type: string
          description: The MCQ question stem (text of the question)
        options:
          type: array
          items: { type: string }
          description: Answer choices (A, B, C, …) for the original MCQ
        learningObjectives:
          type: array
          items: { type: string }
          description: Optional; for alignment with course objectives
      required: [question]

    AssessmentTransformResponse:
      type: object
      properties:
        openEndedQuestion:
          type: string
          description: The transformed open-ended prompt for students
        rubric:
          type: string
          description: Markdown or text rubric; structured rubricCriteria may be added later
        expectedResponseOutline:
          type: string
          description: Rationale or expected response outline for grading
        misconceptions:
          type: array
          items: { type: string }
          description: Common misconceptions / partial-credit guidance (optional)
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
          description: OER sources used to ground the transformation

    Citation:
      type: object
      properties:
        title:
          type: string
          description: Title of the cited source (e.g. book title)
        page:
          type: string
          description: Page reference within the source (optional; may not be available for all OER formats)
        chapter:
          type: string
          description: Chapter where the cited content appears
        section:
          type: string
          description: Section within the chapter (if available)
        snippet:
          type: string
          maxLength: 50
          description: |
            Preview of the retrieved passage (capped at 50). Whitespace is collapsed and ASCII control
            characters removed to reduce PDF-extract noise. Same index as ``<grounded ref="i">``.
      required:
        - title
        - chapter
        - snippet

    Template:
      type: object
      properties:
        id:
          type: string
          description: Stable identifier for this template
        name:
          type: string
          description: Display name (e.g. "Lecture scaffolding")
        description:
          type: string
          description: Short description for the dashboard UI

    TemplateList:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: "#/components/schemas/Template"
          description: |
            Templates for the requested workflow only. Each ``id`` is kebab-case and stable for API use.

    LogEntry:
      type: object
      description: |
        Internal audit log entry. Logged by the Engine on every generate/regenerate
        call. Not exposed as a public API endpoint; documented here for schema reference.
      properties:
        transactionTime:
          type: string
          format: date-time
          description: When the transaction occurred (ISO 8601)
        workflowType:
          type: string
          description: Which workflow was invoked (e.g. "lesson_outline", "assessment")
        success:
          type: boolean
          description: Whether the generation succeeded
        request:
          type: object
          description: Copy of the original request payload (for audit trail)

    TelemetryEvent:
      type: object
      description: |
        Operational telemetry event. Used internally for performance monitoring.
        No public endpoint currently; documented here for schema reference.
      properties:
        eventType:
          type: string
          description: Type of event (e.g. "generate", "retrieve", "error"); no PII
        latencyMs:
          type: number
          description: Request latency in milliseconds
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (ISO 8601)

    Error:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message